requirements.txt


torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
watchdog>=2.1.6
opencv-python>=4.5.5
Pillow>=9.1.0
numpy>=1.21.0
tqdm>=4.64.0